meson: migrate ChangeLog generator to a proper build target
We still run the same script. But now it is run at build time, not
configure time, and Meson checks that it successfully ran -- it
really, really, really, should.

That is, assuming that git is installed and this is a git repo. If not,
we can simply refrain from running anything at all, instead of checking
the return code of gen_changelog.sh