don't hardcode lua version in lua package name
This fails on Fedora Linux, since our Lua 5.4 package is just called `lua`, not `lua53`. (It wouldn't work for our compat package for older versions either, because pkg-config for that is installed as e.g. `lua-5.1`.)

This _should_ work since `req_version` is still there checking the actual version.

If it is necessary to search for the package by this name for some distros, please at least make that a special-case check in addition to the "nameversion" one. Thanks!