Do not truncate socket path
If the length of the socket path is too long to fit in the socket
structure, then we should fail, instead of possibly creating a socket
with an unexpected name, possibly in the the wrong directory.

Limit the socket path to one less than the maximum possible length, so
that the path is always followed by terminating '\0'. This matches
the recommendation in unix(7).