Uses the "::" scope resolution operator for static class methods
This necessary because you can't call a non-static method on the current object
directly from a static method.  The "::" forces the method call to resolve to the
flat C API instead of the C++ class API.  The real fix is to switch to the
`fd->method()` calling pattern, but this is a smaller change for now, and the
calling pattern transition will happen later.