Adjusts filedata.h to create both a C++ API and a C API.
Adds `static` specifier to FileData class functions that don't receive a `FileData` argument
TODO: A future cleanup should get rid of all of the "_list" functions that are basically a foreach.

Adds private declarations for methods marked "_unused"
TODO: Should probably drop these, but that's also not in-scope for this refactor.