Allows `FileDataRef` to skip the initial `file_data_ref` on the `FileData` in question
This allows it to be used in cases where a `FileData` is owned by the scope in which
the `FileDataRef` is being used.  Otherwise, `FileDataRef` would only be able to
prevent a `FileData` from being cleaned up unexpectedly, but would not able to clean
up a `FileData` that we are responsible for.